backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "pytest>=8.3.5",
    "sqlmodel>=0.0.24",
]
